requirements.txt
absl-py==2.3.1
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
astunparse==1.6.3
blinker==1.9.0
cachetools==6.2.2
catboost==1.2.8
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.73.0
debugpy==1.8.17
decorator==5.2.1
docker==7.1.0
docopt==0.6.2
executing==2.2.1
fastapi==0.124.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.61.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
huey==2.5.5
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.7.0
jupyter_core==5.9.1
kaleido==1.2.0
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
logistro==2.0.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
namex==0.1.0
narwhals==2.13.0
nest-asyncio==1.6.0
numpy==2.3.5
opentelemetry-api==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
opt_einsum==3.4.0
optree==0.18.0
optuna==4.6.0
optuna-integration==4.6.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
pipreqs==0.4.13
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
sqlparse==0.5.4
stack-data==0.6.3
starlette==0.50.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.38.0
waitress==3.0.2
wcwidth==0.2.14
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
yarg==0.1.10
zipp==3.23.0
